:fire:

+++ jenkins-init/automation-scripts/generate_key.sh
@@ -3,9 +3,10 @@
 
 ## Note
 
-
-sudo rm ./id_rsa
-sudo rm ./id_rsa.pub
+if [ -f "./id_rsa" ]; then
+    sudo rm ./id_rsa
+    sudo rm ./id_rsa.pub
+fi
 
 ssh-keygen -f ./id_rsa -t rsa -b 4096 -q -N ""
